refactor(web): split useAssistantSyncStream into domain-scoped bus subscriptions (LUM-1992)
Closes LUM-1992

Replaces the monolithic 236-line useAssistantSyncStream hook with three
domain-scoped bus subscriptions:

- useConversationSync (domains/conversations/) — conversation list
  debouncing, per-row metadata refresh, SSE reconnect refetch
- useFeatureFlagBusSync (lib/feature-flags/) — flag cache invalidation
  on sync_changed tags and SSE reconnect
- useAssistantResourceSync (hooks/) — simple tag→queryKey invalidations
  for avatar, identity, config, sounds, schedules, and home-feed events

Each domain now owns its own sync event routing logic instead of all 8+
sync tags routing through a central hub-and-spoke handler. All 22 tests
redistributed across 3 domain-scoped test files. Zero behavioral changes.
