+++ apps/web/src/root-layout.tsx
@@ -10,7 +10,9 @@ import {
 } from "@/domains/chat/hooks/use-assistant-lifecycle";
 import { useAuthStore } from "@/stores/auth-store";
 import { useEnvironmentStore } from "@/lib/environment/environment-store";
-import { useAssistantSyncStream } from "@/hooks/use-assistant-sync-stream";
+import { useAssistantResourceSync } from "@/hooks/use-assistant-resource-sync";
+import { useConversationSync } from "@/domains/conversations/use-conversation-sync";
+import { useFeatureFlagBusSync } from "@/lib/feature-flags/use-feature-flag-bus-sync";
 import { useClientFeatureFlagSync } from "@/lib/feature-flags/use-client-feature-flag-sync";
 import { useAssistantFeatureFlagSync } from "@/lib/feature-flags/use-assistant-feature-flag-sync";
 
@@ -81,10 +83,10 @@ export function RootLayout() {
   });
 
   useAssistantFeatureFlagSync(hasPlatformSession ? lifecycle.assistantId : null);
-  useAssistantSyncStream(
-    lifecycle.assistantId,
-    lifecycle.assistantState.kind === "active",
-  );
+  const isAssistantActive = lifecycle.assistantState.kind === "active";
+  useAssistantResourceSync(lifecycle.assistantId, isAssistantActive);
+  useConversationSync(lifecycle.assistantId, isAssistantActive);
+  useFeatureFlagBusSync(lifecycle.assistantId, isAssistantActive);
 
   useEventBusInit({
     assistantId: lifecycle.assistantId,
